perf: don't load xml formatter on home page
