+++ specifyweb/frontend/js_src/lib/utils/ajax/__tests__/response.test.ts
@@ -1,5 +1,5 @@
 import { handleAjaxError } from '../../../components/Errors/FormatError';
-import { xmlToString } from '../../../components/Syncer/xmlUtils';
+import { xmlToString } from '../../../components/Syncer/xmlToString';
 import { Http, httpCodeToErrorMessage } from '../definitions';
 import { handleAjaxResponse } from '../response';
